requirements.txt
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
reportlab>=4.0.0
Pillow>=10.0.0
yt-dlp>=2024.1.0
